Added directory for static files. Fixed image routes
Please note that this is a band-aid solution for the faulty image paths.

The robust, and correct solution is to change the image paths in the
database before you retrieve them, so that we don't need to parse for
the true image path. To get this working, simply place your images in
the newly created 'uploads' directory.